systematic feature engineering testing; documenting; convert feature engineering into different scripts to make more modular
