+++ src/footai/main.py
@@ -10,7 +10,8 @@ def main():
     parser = create_parser()
     args = parser.parse_args()
     if args.elo_transfer or args.multi_division: args.multi_season=True
-    if args.tier: args.multi_division=False
+    if not hasattr(args, 'tier') : args.tier=None
+    if args.tier :args.multi_division=False
     if args.verbose: print("Running the code with args:", args)
     args.countries = parse_countries(args.countries)  
     if args.division:
